% Generated by roxygen2 (4.0.0): do not edit by hand
\name{FUN}
\alias{FUN}
\title{Fun function}
\usage{
FUN(method, method2, x)
}
\arguments{
\item{method}{one of: \cr
"method1" - very long text here \cr
"method2" - very long text here \cr
"method3" - very long text here \cr
"method4" - very long text here \cr
"method5" - very long text here \cr
"method6" - very long text here \cr
"method7" - very long text here \cr
"method8" - very long text here \cr
"method9" - very long text here \cr
"method10" - very long text here \cr}

\item{method2}{one of:
\itemize{
  \item method1 - very long text here
  \item method2 - very long text here
  \item method3 - very long text here
  \item method4 - very long text here
  \item method5 - very long text here
  \item method6 - very long text here
  \item method7 - very long text here
  \item method8 - very long text here
  \item method9 - very long text here
  \item method10 - very long text here
}}

\item{x}{Yipee}
}
\value{
returns nothing good
}
\description{
I'm smiling already.
}
\examples{
FUN(1, 2, 3)
}
\references{
\url{http://stackoverflow.com/q/18835365/1000343}
}

