% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kill_bot}
\alias{kill_bot}
\title{Curly Single Quote Demo}
\usage{
kill_bot(out = "Bot destroyed!")
}
\arguments{
\item{out}{A message to spit out.}
}
\description{
This demos a problem with curly single quotes that can
not be cut and pasted into the console.
}
\examples{
kill_bot()
kill_bot("Robot missed")
kill_bot("[REGEX](with)[SINGLE](quote)'")
}

