% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paste_op.R
\name{*.character}
\alias{*}
\alias{*.character}
\title{Paste Strings.}
\usage{
\method{*}{character}(x, y)

"*"(x, y)
}
\arguments{
\item{x}{A character string that is to be commented.}

\item{y}{A character string (the comment).}
}
\description{
This operator allows you to paste character strings.
}
\examples{
a <- "The first character string"
b <- "The and the second"

(out <- a*b)
}

