% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SOdemoing-package.R
\docType{data}
\name{stole_this_data}
\alias{stole_this_data}
\title{mtcars Reborn}
\format{A data frame with 32 rows and 11 variables}
\usage{
data(stole_this_data)
}
\description{
A dataset containing car info.
}
\details{
\itemize{
  \item mpg. gas
  \item cyl. cylinders
  \item disp. not a mechanic
  \item hp. horse power
  \item drat. Hmm.
  \item wt. Weight maybe?
  \item qsec. Just a quick second.
  \item vs. Versus :-)
  \item am. Not about the radio I'm guessing.
  \item gear. Number of gears.
  \item carb. Number of carbohydrates consumed or maybe carburetor.
}
}
\references{
\code{\link[datasets]{mtcars}}
}
\keyword{datasets}

