% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stuff}
\alias{stuff}
\alias{stuff.data.frame}
\alias{stuff.default}
\title{stuff Methods}
\usage{
stuff(x, ...)

\method{stuff}{data.frame}(x, ...)

\method{stuff}{default}(x, ...)
}
\arguments{
\item{x}{An object}

\item{\ldots}{They make me put these in here}
}
\description{
It does some random stuff

\code{stuff.data.frame} - data.frame method for \code{stuff} used to
convert to stuff.

\code{stuff.default} - default method for \code{stuff} used to
convert to stuff.
}

